Simplify ECB currency exchange plugin
